% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_wrappers.R
\name{GRF_SurvivalForestWrapper}
\alias{GRF_SurvivalForestWrapper}
\title{GRF Survival Forest Model}
\description{
Implements a wrapper for generalized random forest survival models using the \pkg{grf} package.
}
\section{Super class}{
\code{\link[confsurv:SurvivalModelWrapper]{confsurv::SurvivalModelWrapper}} -> \code{GRF_SurvivalForestWrapper}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GRF_SurvivalForestWrapper-fit}{\code{GRF_SurvivalForestWrapper$fit()}}
\item \href{#method-GRF_SurvivalForestWrapper-predict}{\code{GRF_SurvivalForestWrapper$predict()}}
\item \href{#method-GRF_SurvivalForestWrapper-clone}{\code{GRF_SurvivalForestWrapper$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="confsurv" data-topic="SurvivalModelWrapper" data-id="initialize"><a href='../../confsurv/html/SurvivalModelWrapper.html#method-SurvivalModelWrapper-initialize'><code>confsurv::SurvivalModelWrapper$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="confsurv" data-topic="SurvivalModelWrapper" data-id="parse_formula"><a href='../../confsurv/html/SurvivalModelWrapper.html#method-SurvivalModelWrapper-parse_formula'><code>confsurv::SurvivalModelWrapper$parse_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="confsurv" data-topic="SurvivalModelWrapper" data-id="predict_interpolate"><a href='../../confsurv/html/SurvivalModelWrapper.html#method-SurvivalModelWrapper-predict_interpolate'><code>confsurv::SurvivalModelWrapper$predict_interpolate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="confsurv" data-topic="SurvivalModelWrapper" data-id="predict_quantiles"><a href='../../confsurv/html/SurvivalModelWrapper.html#method-SurvivalModelWrapper-predict_quantiles'><code>confsurv::SurvivalModelWrapper$predict_quantiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="confsurv" data-topic="SurvivalModelWrapper" data-id="select_columns"><a href='../../confsurv/html/SurvivalModelWrapper.html#method-SurvivalModelWrapper-select_columns'><code>confsurv::SurvivalModelWrapper$select_columns()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRF_SurvivalForestWrapper-fit"></a>}}
\if{latex}{\out{\hypertarget{method-GRF_SurvivalForestWrapper-fit}{}}}
\subsection{Method \code{fit()}}{
Fit a generalized random forest survival model using the \pkg{grf} package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRF_SurvivalForestWrapper$fit(formula, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{A survival formula of the form \code{Surv(time, status) ~ predictors}.}

\item{\code{data}}{A data.frame containing columns for time, status, and covariates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRF_SurvivalForestWrapper-predict"></a>}}
\if{latex}{\out{\hypertarget{method-GRF_SurvivalForestWrapper-predict}{}}}
\subsection{Method \code{predict()}}{
Predicts survival probabilities at specified time points for new data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRF_SurvivalForestWrapper$predict(new_data, time.points = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{A data.frame of new observations with the same structure as the training data.}

\item{\code{time.points}}{Optional numeric vector of time points at which to compute predictions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with:
\describe{
\item{\code{predictions}}{Matrix of survival probabilities (rows = individuals, cols = time points).}
\item{\code{time.points}}{Vector of time points used in prediction.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRF_SurvivalForestWrapper-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GRF_SurvivalForestWrapper-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRF_SurvivalForestWrapper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
