% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{interpret_screening_rule}
\alias{interpret_screening_rule}
\title{Format Screening Rule as String}
\usage{
interpret_screening_rule(time, prob, crit)
}
\arguments{
\item{time}{Screening time.}

\item{prob}{Survival probability threshold.}

\item{crit}{Screening direction (\code{"low risk"} or \code{"high risk"}).}
}
\value{
A formatted string describing the rule.
}
\description{
Format Screening Rule as String
}
