% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_wrappers.R
\name{randomForestSRC_SurvivalWrapper}
\alias{randomForestSRC_SurvivalWrapper}
\title{RandomForestSRC Survival Wrapper}
\description{
Wrapper for the \code{randomForestSRC::rfsrc()} survival model.
}
\section{Super class}{
\code{\link[confsurv:SurvivalModelWrapper]{confsurv::SurvivalModelWrapper}} -> \code{randomForestSRC_SurvivalWrapper}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-randomForestSRC_SurvivalWrapper-fit}{\code{randomForestSRC_SurvivalWrapper$fit()}}
\item \href{#method-randomForestSRC_SurvivalWrapper-predict}{\code{randomForestSRC_SurvivalWrapper$predict()}}
\item \href{#method-randomForestSRC_SurvivalWrapper-clone}{\code{randomForestSRC_SurvivalWrapper$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="confsurv" data-topic="SurvivalModelWrapper" data-id="initialize"><a href='../../confsurv/html/SurvivalModelWrapper.html#method-SurvivalModelWrapper-initialize'><code>confsurv::SurvivalModelWrapper$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="confsurv" data-topic="SurvivalModelWrapper" data-id="parse_formula"><a href='../../confsurv/html/SurvivalModelWrapper.html#method-SurvivalModelWrapper-parse_formula'><code>confsurv::SurvivalModelWrapper$parse_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="confsurv" data-topic="SurvivalModelWrapper" data-id="predict_interpolate"><a href='../../confsurv/html/SurvivalModelWrapper.html#method-SurvivalModelWrapper-predict_interpolate'><code>confsurv::SurvivalModelWrapper$predict_interpolate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="confsurv" data-topic="SurvivalModelWrapper" data-id="predict_quantiles"><a href='../../confsurv/html/SurvivalModelWrapper.html#method-SurvivalModelWrapper-predict_quantiles'><code>confsurv::SurvivalModelWrapper$predict_quantiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="confsurv" data-topic="SurvivalModelWrapper" data-id="select_columns"><a href='../../confsurv/html/SurvivalModelWrapper.html#method-SurvivalModelWrapper-select_columns'><code>confsurv::SurvivalModelWrapper$select_columns()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-randomForestSRC_SurvivalWrapper-fit"></a>}}
\if{latex}{\out{\hypertarget{method-randomForestSRC_SurvivalWrapper-fit}{}}}
\subsection{Method \code{fit()}}{
Fits a survival forest model using the \pkg{randomForestSRC} package.
This method parses the survival formula and fits a random survival forest to the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{randomForestSRC_SurvivalWrapper$fit(formula, data, ntree = 100, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{A survival formula, typically \code{Surv(time, status) ~ predictors}.}

\item{\code{data}}{A data.frame containing time, status, and covariates.}

\item{\code{ntree}}{Number of trees to grow in the forest (default = 100).}

\item{\code{...}}{Additional arguments passed to \code{randomForestSRC::rfsrc()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-randomForestSRC_SurvivalWrapper-predict"></a>}}
\if{latex}{\out{\hypertarget{method-randomForestSRC_SurvivalWrapper-predict}{}}}
\subsection{Method \code{predict()}}{
Predicts survival probabilities at specified time points for new data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{randomForestSRC_SurvivalWrapper$predict(new_data, time.points = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{A data.frame of new observations with the same structure as the training data.}

\item{\code{time.points}}{Optional numeric vector of time points at which to compute predictions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with:
\describe{
\item{\code{predictions}}{Matrix of survival probabilities (rows = individuals, cols = time points).}
\item{\code{time.points}}{Vector of time points used in prediction.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-randomForestSRC_SurvivalWrapper-clone"></a>}}
\if{latex}{\out{\hypertarget{method-randomForestSRC_SurvivalWrapper-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{randomForestSRC_SurvivalWrapper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
