% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\docType{class}
\name{SurvivalDataGenerator}
\alias{SurvivalDataGenerator}
\title{SurvivalDataGenerator: Synthetic Survival Data Generator}
\value{
An object of class \code{SurvivalDataGenerator}, with a \verb{$sample()} method that returns
a data.frame including: \code{event_time}, \code{censoring_time}, \code{time}, \code{status}, and covariates \verb{X1, X2, ..., Xp}.
}
\description{
An R6 class for generating synthetic survival data using user-defined covariate generators,
survival time models, and censoring mechanisms. The class composes these components to simulate
realistic right-censored survival datasets.
}
\examples{
cov_gen <- function(n) matrix(rnorm(n * 2), nrow = n)
surv_gen <- ExponentialDistribution$new(function(X) rep(0.1, nrow(X)))
cens_gen <- ExponentialDistribution$new(function(X) rep(0.05, nrow(X)))
gen <- SurvivalDataGenerator$new(cov_gen, surv_gen, cens_gen)
data <- gen$sample(100)
head(data)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{covariate_generator}}{Function taking an integer \code{num_samples} and returning a matrix or data.frame.}

\item{\code{survival}}{An object with a \code{sample(X)} method to simulate event times based on covariates.}

\item{\code{censoring}}{An object with a \code{sample(X)} method to generate censoring times based on covariates.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SurvivalDataGenerator-new}{\code{SurvivalDataGenerator$new()}}
\item \href{#method-SurvivalDataGenerator-sample}{\code{SurvivalDataGenerator$sample()}}
\item \href{#method-SurvivalDataGenerator-clone}{\code{SurvivalDataGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurvivalDataGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-SurvivalDataGenerator-new}{}}}
\subsection{Method \code{new()}}{
Constructor for SurvivalDataGenerator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalDataGenerator$new(
  covariate_generator,
  survival_generator,
  censoring_generator
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariate_generator}}{Function to generate covariates.}

\item{\code{survival_generator}}{Object for generating event times.}

\item{\code{censoring_generator}}{Object for generating censoring times.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurvivalDataGenerator-sample"></a>}}
\if{latex}{\out{\hypertarget{method-SurvivalDataGenerator-sample}{}}}
\subsection{Method \code{sample()}}{
Generate a dataset of survival data with censoring.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalDataGenerator$sample(num_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_samples}}{Number of observations to simulate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurvivalDataGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SurvivalDataGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurvivalDataGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
